refactor: UI 클래스를 InputView와 OutputView 클래스로 분리
UI 클래스가 입력과 출력이라는 두 가지 책임을 모두 가지고 있어 이를 분리함
