+++ README.md
@@ -86,7 +86,7 @@ public int validateRoundLimit(String input)
 - [x]  input이 숫자 형태가 아니거나, 양의 정수가 아니거나, 100 초과인 경우 예외 발생
 - [x]  예외가 발생하지 않았다면 정수로 변환된 input을 리턴
 
-## UI 클래스
+## InputView 클래스
 
 public String[] getCarNames()
 : \: 경주할 자동차 이름을 입력받는 메소드.
@@ -103,6 +103,8 @@ public int getRoundLimit()
 - [x]  `camp.nextstep.edu.missionutils.Console`의`readLine()`을 활용해 시도할 횟수 입력받기
 - [x]  Validate.validateRoundLimit() 호출해 입력받은 값 검증
 - [x]  입력받은 값에 문제가 없다면 해당 입력 값을 정수로 변환한 값을 리턴
+
+## OutputView 클래스
 
 public void printRoundResult(boolean isFirstRound, List\<Car> participants)
 : \: 각 라운드의 결과를 출력하는 메소드.
