+++ src/main/java/racingcar/Application.java
@@ -1,18 +1,16 @@
 package racingcar;
 
 import racingcar.domain.RacingGame;
+import racingcar.view.InputView;
 
 public class Application {
     public static void main(String[] args) {
-        UI ui = new UI();
-
-        String[] carNames = ui.getCarNames();
-        int limit = ui.getRoundLimit();
+        InputView inputView = new InputView();
+        String[] carNames = inputView.getCarNames();
+        int limit = inputView.getRoundLimit();
 
         RacingGame racingGame = new RacingGame(limit);
-
         racingGame.setParticipants(carNames);
-
         racingGame.playGame();
     }
 }
